HYP-28: Move management endpoints from projects app to manage app
Also fixed issues with some malformed logging statements, added necessary SciAuthZ permissions checks to some functions. Combined remaining functions into a projects/api.py file since these functions aren't really "views".
