+++ app/projects/api.py
@@ -1,55 +1,281 @@
-import json
-import logging
-import os
-import shutil
-import zipfile
-import uuid
-
 from copy import copy
 from datetime import datetime
-
-import requests
+import json
+import logging
 
 from django.conf import settings
+from django.contrib import messages
+from django.contrib.auth.models import User
 from django.core import exceptions
-from django.http import HttpResponse
+from django.core.exceptions import ObjectDoesNotExist
 from django.http import JsonResponse
+from django.http import HttpResponse
 from django.http import QueryDict
-from django.shortcuts import redirect
 from django.shortcuts import get_object_or_404
+from django.shortcuts import redirect
 
-from pyauth0jwt.auth0authenticate import user_auth_and_jwt
-
-from hypatio.sciauthz_services import SciAuthZ
-from hypatio.file_services import get_download_url
+from contact.views import email_send
 from hypatio import file_services as fileservice
+from hypatio.file_services import get_download_url
+from hypatio.sciauthz_services import SciAuthZ
+from projects.templatetags import projects_extras
 
+from projects.models import ChallengeTask
+from projects.models import ChallengeTaskSubmission
 from projects.models import DataProject
 from projects.models import HostedFile
 from projects.models import HostedFileDownload
-from projects.models import ChallengeTask
-from projects.models import ChallengeTaskSubmission
-from projects.models import TeamSubmissionsDownload
 from projects.models import Participant
 from projects.models import Team
 
-from projects.templatetags import projects_extras
-
-from contact.views import email_send
-
-from contact.views import email_send
+from pyauth0jwt.auth0authenticate import user_auth_and_jwt
 
 logger = logging.getLogger(__name__)
 
+@user_auth_and_jwt
+def finalize_team(request):
+    """
+    An HTTP POST endpoint for team leaders to mark their team as ready and send an email notification to contest managers.
+    """
+
+    project_key = request.POST.get("project_key")
+    team = request.POST.get("team")
+
+    project = DataProject.objects.get(project_key=project_key)
+    team = Team.objects.get(team_leader__email=team, data_project=project)
+
+    if request.user.email != team.team_leader.email:
+        logger.debug(
+            "User {email} is not a team leader.".format(
+                email=request.user.email
+            )
+        )
+        return HttpResponse("Error: permissions.", status=403)
+
+    team.status = 'Ready'
+    team.save()
+
+    # TODO Eventually this should be replaced by checking the dataproject's leader
+    contest_managers = ['ouzuner@gmu.edu', 'filannim@csail.mit.edu', 'stubbs@simmons.edu']
+
+    context = {'team_leader': team,
+               'project': project_key,
+               'site_url': settings.SITE_URL}
+
+    email_success = email_send(subject='DBMI Portal - Finalized Team',
+                               recipients=contest_managers,
+                               email_template='email_finalized_team_notification',
+                               extra=context)
+
+    return HttpResponse(200)
+
+@user_auth_and_jwt
+def approve_team_join(request):
+    """
+    An HTTP POST endpoint for team leaders to approve requests from others to join their team.
+    """
+
+    project_key = request.POST.get("project_key")
+    participant_email = request.POST.get("participant")
+
+    project = DataProject.objects.get(project_key=project_key)
+
+    try:
+        team = Team.objects.get(
+            team_leader=request.user,
+            data_project=project
+        )
+    except ObjectDoesNotExist:
+        team = None
+
+    if request.user.email != team.team_leader.email:
+        logger.debug(
+            "User {email} is not a team leader.".format(
+                email=request.user.email
+            )
+        )
+        return HttpResponse("Error: permissions.", status=403)
+
+    try:
+        participant_user = User.objects.get(email=participant_email)
+        participant = Participant.objects.get(
+            user=participant_user,
+            data_challenge=project
+        )
+    except ObjectDoesNotExist:
+        participant = None
+
+    participant.assign_approved(team)
+    participant.save()
+
+    return HttpResponse(200)
+
+@user_auth_and_jwt
+def reject_team_join(request):
+    """
+    An HTTP POST endpoint for team leaders to reject requests from others to join their team.
+    """
+
+    project_key = request.POST.get("project_key")
+    participant_email = request.POST.get("participant")
+
+    project = DataProject.objects.get(project_key=project_key)
+
+    try:
+        participant_user = User.objects.get(email=participant_email)
+        participant = Participant.objects.get(
+            user=participant_user,
+            data_challenge=project
+        )
+    except ObjectDoesNotExist:
+        logger.debug('Participant not found.')
+        return HttpResponse('Error.', status=404)
+
+    if request.user.email != participant.team.team_leader.email:
+        logger.debug(
+            "User {email} is not a team leader.".format(
+                email=request.user.email
+            )
+        )
+        return HttpResponse("Error: permissions.", status=403)
+
+    participant.delete()
+
+    return HttpResponse(200)
+
+@user_auth_and_jwt
+def leave_team(request):
+    """
+    An HTTP POST endpoint for removing the participant from whatever team they are currently on or have requested to be on.
+    """
+
+    project_key = request.POST.get("project_key", "")
+
+    logger.debug("[HYPATIO][leave_team] User " + request.user.email + " trying to leave their current team for project " + project_key + ".")
+
+    try:
+        project = DataProject.objects.get(project_key=project_key)
+    except ObjectDoesNotExist:
+        logger.error("[HYPATIO][leave_team] DataProject not found for given project_key: " + project_key)
+        return HttpResponse(500)
+
+    # TODO user does not have permissions to remove their view permission (whether or not it exists)
+    # Remove VIEW permissions on the DataProject
+    # sciauthz = SciAuthZ(settings.AUTHZ_BASE, request.COOKIES.get("DBMI_JWT", None), request.user.email)
+    # sciauthz.remove_view_permission(project_key, request.user.email)
+
+    # TODO remove team leader's scireg permissions
+    # ...
+
+    participant = Participant.objects.get(user=request.user, data_challenge=project)
+    participant.team = None
+    participant.pending = False
+    participant.approved = False
+    participant.team_wait_on_leader_email = None
+    participant.team_wait_on_leader = False
+    participant.save()
+
+    return redirect('/projects/' + request.POST.get('project_key') + '/')
+
+@user_auth_and_jwt
+def join_team(request):
+    """
+    An HTTP POST endpoint for a user to try to join a team by entering the team leader's email.
+    """
+
+    project_key = request.POST.get("project_key")
+    project = DataProject.objects.get(project_key=project_key)
+
+    team_leader = request.POST.get("team_leader")
+
+    try:
+        participant = Participant.objects.get(user=request.user, data_challenge=project)
+    except ObjectDoesNotExist:
+        participant = Participant(user=request.user, data_challenge=project)
+        participant.save()
+
+    try:
+        # If this team leader has already created a team, add the person to the team in a pending status
+        team = Team.objects.get(team_leader__email__iexact=team_leader, data_project=project)
+
+        # Only allow a new participant to join a team that is still in a pending or ready state
+        if team.status in ['Pending', 'Ready']:
+            participant.team = team
+            participant.team_pending = True
+            participant.save()
+        # Otherwise, let them know why they can't join the team
+        else:
+            msg = "The team you are trying to join has already been finalized and is not accepting new members. " + \
+                  "If you would like to join this team, please have the team leader contact the challenge " + \
+                  "administrators for help."
+            messages.error(request, msg)
+
+            return redirect('/projects/' + request.POST.get('project_key') + '/')
+    except ObjectDoesNotExist:
+        # If this team leader has not yet created a team, mark the person as waiting
+        participant.team_wait_on_leader_email = team_leader
+        participant.team_wait_on_leader = True
+        participant.save()
+        team = None
+
+    # Send email to team leader informing them of a pending member
+    if team is not None:
+        context = {'member_email': request.user.email,
+                   'project': project,
+                   'site_url': settings.SITE_URL}
+
+        email_success = email_send(subject='DBMI Portal - Pending Member',
+                                   recipients=[team_leader],
+                                   email_template='email_pending_member_notification',
+                                   extra=context)
+
+    logger.debug('[HYPATIO][join_team] - Creating Profile Permissions')
+
+    # Create record to allow leader access to profile.
+    sciauthz = SciAuthZ(settings.AUTHZ_BASE, request.COOKIES.get("DBMI_JWT", None), request.user.email)
+    sciauthz.create_profile_permission(team_leader, project_key)
+
+    return redirect('/projects/' + request.POST.get('project_key') + '/')
+
+@user_auth_and_jwt
+def create_team(request):
+    """
+    An HTTP POST endpoint for users to create a new team and be its team leader.
+    """
+
+    project_key = request.POST.get("project_key")
+    project = DataProject.objects.get(project_key=project_key)
+    new_team = Team.objects.create(team_leader=request.user, data_project=project)
+
+    try:
+        participant = Participant.objects.get(user=request.user, data_challenge=project)
+    except ObjectDoesNotExist:
+        participant = Participant(user=request.user, data_challenge=project)
+        participant.save()
+
+    participant.assign_approved(new_team)
+    participant.save()
+
+    # Find anyone whose waiting on this team leader and link them to the new team.
+    waiting_participants = Participant.objects.filter(
+        team_wait_on_leader_email=request.user.email,
+        data_challenge=project
+    )
+
+    for participant in waiting_participants:
+        participant.assign_pending(new_team)
+        participant.save()
+
+    return redirect('/projects/' + project_key + '/')
 
 @user_auth_and_jwt
 def download_dataset(request):
     """
-    Handles downloads for project level files. Checks that the requesting user
+    An HTTP GET endpoint that handles downloads for project level files. Checks that the requesting user
     has view permissions on the given project before allowing a download.
     """
 
-    logger.debug("[views_files][download_dataset] - Attempting file download.")
+    logger.debug("[download_dataset] - Attempting file download.")
 
     file_uuid = request.GET.get("file_uuid")
     file_to_download = get_object_or_404(HostedFile, uuid=file_uuid)
@@ -57,7 +283,7 @@ def download_dataset(request):
 
     # Check if this file is enabled for download.
     if not projects_extras.is_hostedfile_currently_enabled(file_to_download):
-        logger.debug("[views_files][download_dataset] - File not allowed for download attempted by " + request.user.email)
+        logger.debug("[download_dataset] - File not allowed for download attempted by " + request.user.email)
         return HttpResponse("You do not have access to download this file.", status=403)
 
     # Check for necessary permissions in SciAuthZ.
@@ -65,14 +291,14 @@ def download_dataset(request):
     sciauthz = SciAuthZ(settings.AUTHZ_BASE, user_jwt, request.user.email)
 
     if not sciauthz.user_has_single_permission(project_key, "VIEW"):
-        logger.debug("[views_files][download_dataset] - No Access for user " + request.user.email)
+        logger.debug("[download_dataset] - No Access for user " + request.user.email)
         return HttpResponse("You do not have access to download this file.", status=403)
 
     # Save a record of this person downloading this file.
     HostedFileDownload.objects.create(user=request.user, hosted_file=file_to_download)
 
     s3_filename = file_to_download.file_location + "/" + file_to_download.file_name
-    logger.debug("[views_files][download_dataset] - User " + request.user.email + " is downloading file " + s3_filename + " from bucket " + settings.S3_BUCKET + ".")
+    logger.debug("[download_dataset] - User " + request.user.email + " is downloading file " + s3_filename + " from bucket " + settings.S3_BUCKET + ".")
 
     download_url = get_download_url(s3_filename)
 
@@ -81,158 +307,6 @@ def download_dataset(request):
     response['filename'] = file_to_download.file_name
 
     return response
-
-@user_auth_and_jwt
-def download_team_submissions(request):
-    """
-    Handles downloads of participant submission files. Checks that the requesting user
-    has proper permissions to access this file. Downloads all of the team's submissions
-    into one zip file, with each submission in it's own zip file containing a json file
-    with metadata about the upload.
-    """
-    logger.debug('download_team_submissions: {}'.format(request.method))
-
-    if request.method == "GET":
-
-        project_key = request.GET.get("project_key", "")
-        team_leader_email = request.GET.get("team", "")
-
-        if project_key == "" or team_leader_email == "":
-            return HttpResponse("Not all parameters provided.", status=404)
-
-        # Check Permissions in SciAuthZ
-        user_jwt = request.COOKIES.get("DBMI_JWT", None)
-        sciauthz = SciAuthZ(settings.AUTHZ_BASE, user_jwt, request.user.email)
-        is_manager = sciauthz.user_has_manage_permission(project_key)
-
-        if not is_manager:
-            logger.debug("[views_files][download_team_submissions] - No Access for user " + request.user.email)
-            return HttpResponse("You do not have access to download this file.", status=403)
-
-        project = get_object_or_404(DataProject, project_key=project_key)
-        team = get_object_or_404(Team, data_project=project, team_leader__email=team_leader_email)
-        team_participants = team.participant_set.all()
-
-        logger.debug('Building a zip file containing the submissions from team {team}.'.format(
-            team=team_leader_email
-        ))
-
-        # Get all of the participant submission records belonging to this team, ordered by upload date
-        team_submissions = team.get_submissions().order_by('upload_date')
-
-        # Save a record of this person downloading this team's submissions file
-        download_record = TeamSubmissionsDownload.objects.create(
-            user=request.user,
-            team=team
-        )
-
-        # Once the record is created and a pk assigned, update with the specific submissions referenced
-        download_record.participant_submissions = team_submissions
-        download_record.save()
-
-        # Create a dictory to hold the zipped submissions using a guid to keep it isolated from other requests
-        zipped_submissions_directory = "/tmp/zipped_submissions-" + str(uuid.uuid4())
-        if not os.path.exists(zipped_submissions_directory):
-            os.makedirs(zipped_submissions_directory)
-
-        # A list of file paths to each task's zip file
-        zipped_tasks_filepath = []
-
-        # Create subdirectories for each challenge task
-        for task in project.challengetask_set.all():
-
-            # A list of file paths to each submission zip file for this task
-            zipped_task_submission_filepaths = []
-
-            # Get the submissions for this task submitted by the team.
-            submissions = ChallengeTaskSubmission.objects.filter(
-                challenge_task=task,
-                participant__in=team.participant_set.all(),
-                deleted=False
-            )
-
-            if submissions.count() > 0:
-
-                for i, submission in enumerate(submissions):
-                    submission_number = i + 1
-                    submission_date_string = datetime.strftime(submission.upload_date, "%Y%m%d_%H%M")
-
-                    # Create a working directory to hold the files specific to this submission that need to be zipped together
-                    working_directory = "/tmp/working_dir-" + str(uuid.uuid4())
-                    if not os.path.exists(working_directory):
-                        os.makedirs(working_directory)
-
-                    # Create a json file with the submission info string
-                    info_file_name = "submission_info.json"
-                    with open(working_directory + "/" + info_file_name, mode="w") as f:
-                        f.write(submission.submission_info)
-
-                    # Get the submission file's byte contents from S3
-                    submission_file_download_url = fileservice.get_fileservice_download_url(request, submission.uuid)
-                    submission_file_request = requests.get(submission_file_download_url)
-
-                    # Write the bytes to a zip file
-                    submission_file_name = "submission_file.zip"
-                    if submission_file_request.status_code == 200:
-                        with open(working_directory + "/" + submission_file_name, mode="wb") as f:
-                            f.write(submission_file_request.content)
-                    else:
-                        logger.error("[views_files][download_team_submissions] - Participant submission {uuid} file could not be pulled from S3.".format(
-                            uuid=submission.uuid
-                        ))
-                        return HttpResponse("Error getting files", status=404)
-
-                    # Create the zip file and add the files to it
-                    zip_file_path = zipped_submissions_directory + "/" + submission.challenge_task.title + "_" + str(submission_number) + "_" + submission_date_string + ".zip"
-                    with zipfile.ZipFile(zip_file_path, mode="w") as zf:
-                        zf.write(working_directory + "/" + info_file_name, arcname=info_file_name)
-                        zf.write(working_directory + "/" + submission_file_name, arcname=submission_file_name)
-
-                    # Add the zipfile to the list of zip files
-                    zipped_task_submission_filepaths.append(zip_file_path)
-
-                    # Delete the working directory
-                    shutil.rmtree(working_directory)
-
-                # Create a directory to store the encompassing task's zip file using a guid to keep it isolated from other requests
-                task_zip_file_directory = "/tmp/task_zip-" + str(uuid.uuid4())
-                if not os.path.exists(task_zip_file_directory):
-                    os.makedirs(task_zip_file_directory)
-
-                # Create the encompassing task's zip file
-                task_zip_file_name = task.title + ".zip"
-                task_zip_file_path = os.path.join(task_zip_file_directory, task_zip_file_name)
-                with zipfile.ZipFile(task_zip_file_path, mode="w") as zf:
-                    for submission_zip in zipped_task_submission_filepaths:
-                        zf.write(submission_zip, arcname=os.path.basename(submission_zip))
-
-                # Add the zipfile to the list of zip files
-                zipped_tasks_filepath.append(task_zip_file_path)
-
-        # Create a directory to store the encompassing zip file using a guid to keep it isolated from other requests
-        final_zip_file_directory = "/tmp/final_zip-" + str(uuid.uuid4())
-        if not os.path.exists(final_zip_file_directory):
-            os.makedirs(final_zip_file_directory)
-
-        # Combine all the zipped tasks into one file zip file
-        final_zip_file_name = project_key + "_submissions_" + team_leader_email.replace('@', '-at-') + ".zip"
-        final_zip_file_path = os.path.join(final_zip_file_directory, final_zip_file_name)
-        with zipfile.ZipFile(final_zip_file_path, mode="w") as zf:
-            for task_zip in zipped_tasks_filepath:
-                zf.write(task_zip, arcname=os.path.basename(task_zip))
-
-        # Prepare the zip file to be served
-        final_zip_file = open(final_zip_file_path, 'rb')
-        response = HttpResponse(final_zip_file, content_type='application/force-download')
-        response['Content-Disposition'] = 'attachment; filename="%s"' % final_zip_file_name
-
-        # Delete all the zip files from disk storage
-        shutil.rmtree(final_zip_file_directory)
-        shutil.rmtree(zipped_submissions_directory)
-        for path in zipped_tasks_filepath:
-            shutil.rmtree(os.path.dirname(path))
-
-        return response
 
 @user_auth_and_jwt
 def upload_challengetasksubmission_file(request):
@@ -260,7 +334,7 @@ def upload_challengetasksubmission_file(request):
         sciauthz = SciAuthZ(settings.AUTHZ_BASE, user_jwt, request.user.email)
 
         if not sciauthz.user_has_single_permission(project_key, "VIEW"):
-            logger.debug("[views_files][upload_challengetasksubmission_file] - No Access for user " + request.user.email)
+            logger.debug("[upload_challengetasksubmission_file] - No Access for user " + request.user.email)
             return HttpResponse("You do not have access to upload this file.", status=403)
 
         if filename.split(".")[-1] != "zip":
@@ -275,7 +349,7 @@ def upload_challengetasksubmission_file(request):
 
         # Only allow a submission if the task is still open.
         if not projects_extras.is_challengetask_currently_enabled(task):
-            logger.error("[views_files][upload_challengetasksubmission_file] - User " + request.user.email + " is trying to submit task " + task.title + " after close time.")
+            logger.error("[upload_challengetasksubmission_file] - User " + request.user.email + " is trying to submit task " + task.title + " after close time.")
             return HttpResponse("This task is no longer open for submissions", status=400)
 
         # Prepare the metadata.
@@ -394,8 +468,8 @@ def upload_challengetasksubmission_file(request):
 @user_auth_and_jwt
 def delete_challengetasksubmission(request):
     """
-    Marks a ChallengeTaskSubmission as deleted so it will not be counted against their
-    total submission count for a contest.
+    An HTTP POST endpoint that marks a ChallengeTaskSubmission as deleted so
+    it will not be counted against their total submission count for a contest.
     """
 
     if request.method == 'POST':
@@ -412,13 +486,13 @@ def delete_challengetasksubmission(request):
 
         if not sciauthz.user_has_single_permission(project.project_key, "VIEW"):
             logger.debug(
-                "[views_files][delete_challengetasksubmission] - No Access for user %s",
+                "[delete_challengetasksubmission] - No Access for user %s",
                 request.user.email
             )
             return HttpResponse("You do not have access to delete this file.", status=403)
 
         logger.debug(
-            '[views_files][delete_challengetasksubmission] - %s is trying to delete submission %s',
+            '[delete_challengetasksubmission] - %s is trying to delete submission %s',
             request.user.email,
             submission_uuid
         )
@@ -430,7 +504,7 @@ def delete_challengetasksubmission(request):
         # Check that the user is either the team leader, the original submitter, or a manager
         if not user_is_submitter and not user_is_team_leader and not user_is_manager:
             logger.debug(
-                "[views_files][delete_challengetasksubmission] - No Access for user %s",
+                "[delete_challengetasksubmission] - No Access for user %s",
                 request.user.email
             )
             return HttpResponse("Only the original submitter, team leader, or challenge manager may delete this.", status=403)
